refactor params and allow config of multiple params
