+++ tubul/tubul_params.h
@@ -10,6 +10,9 @@
 
 namespace TU
 {
+    // Reset any parameter configuration
+    void clearParams();
+
     // Passes the configuration of parameters to parse.
     void configParams(std::istream& paramsConfig);
     void configParams(const std::string& paramsConfig);
